[1/n] Editor Scaffolding - Add EditorContainer
